Move shared validations to a concern
